Module arguments: lib
Modules can be just an attribute set. But if you want access to some
arguments you need to change it into a function taking an attribute set
with an ellipsis (that's the "..."). In this case we only match on the
`lib` argument, which gives us access to nixpkgs library functions.

Note that the ellipsis is necessary since arbitrary arguments can be
passed to modules.