Declaring generate.script option
In order for modules to be useful, we need to have options, so let's
declare one. Options are declared by defining an attribute with
`lib.mkOption` under the `options` attribute. Here we're defining the
option `generate.script`.

While there are many attributes to customize options, the most
important one is `type`, which specifies what values are valid for an
option, and how/whether multiple values should be merged together.

In the nixpkgs library there are a number of types available under
`lib.types`. Here we're using the `lines` type, which specifies that:
- Only strings are valid values
- Multiple strings are joined with newlines

Trying to evaluate the `generate.script` option however, we get an error
that the option is used but not defined, indicating that we need to
actually give a value to the option.