Successful evaluation: Assigning a string to generate.script
We can make type checking pass by assigning a string to the option,
giving us our first successful evaluation of the `generate.script`
option.

In this case, we assign a script which calls the Google Maps Static API
to generate a world map, then displaying the result using icat
(image-cat), both of which are helper scripts.