A new list option: Declaring generate.requestParams
Let's introduce another option, generate.requestParams. For its type,
we'll use `listOf <nestedType>`, which is a generic list type where each
element has to match the given nested type. In our case we want `str` to
be the nested type, which is a generic string type.

Note that the difference between `str` and `lines` is in their merging
behavior:
- For `lines`, multiple definitions get merged by concatenation with
  newlines
- For `str`, multiple definitions are not allowed. Which in this case is
  mostly irrelevant however, since we can't really define a list element
  multiple times.