Dependencies between options: Using generate.requestParams
A collection of modules generally only has a single option that is meant
to be evaluated. That's the option that generates the final result we're
interested in, which in our case is `generate.script`.

In order to build up abstractions on that, we have the ability for
options to depend on other options. In this case we want the
`generate.script` option to use the values of `generate.requestParams`.
We can access the values of options by adding the `config` argument to
the argument list at the top and using e.g.
`config.generate.requestParams` to access that options value.

We're then using `lib.concatStringsSep " "` to join each list element
of the option together into an argument list.