Conditional definitions: Introducing map.zoom option
We now introduce the new option `map.zoom` in order to control the zoom
level of the map. We'll use a new type for it, `nullOr <type>`, which
accepts the value `null`, but also values of its argument type. We're
using `null` here to mean an inferred zoom level.

For this option, we'll set a default value which should be used if it's
not defined otherwise, which we can do using `mkOption`'s `default`
argument.

Now we want to use this option to define another element in
`generate.requestParams`, but we only want to add this element if its
value is non-null. We can do this using the `mkIf <condition>
<definition>` function, which only adds a definition if the condition
holds.