Submodule types: Declaring map.markers option
One of the most useful types of the module system is the `submodule`
type. This type allows you to define a nested module system evaluation,
with its own options. Every value of such a type is then interpreted
(by default) as a `config` assignment of the nested module evaluation.

In this case we're defining a `map.markers` option, whose type is a list
of submodules with a nested `location` type, allowing us to define a
list of markers on the map, where each assignment is type checked
according to the submodule.