Nested submodules: Introducing users option
We will now introduce the option `users`, where we make use of a very
useful type, `lib.types.attrsOf <subtype>`. This type lets us specify an
attribute set as a value, where the keys can be arbitrary, but each
value has to conform to the given <subtype>.

In this case, we'll use another submodule as the subtype, one that
allows declaring a departure marker, which notably also makes use of our
`markerType` submodule, giving us a nested structure of submodules.

We're now propagating each of the users marker definitions to the
`map.markers` option.