Using the attribute name in the submodule to define a default label
Let's set a default label by deriving it from the username. By
transforming the submodule's argument into a function, we can access
arguments within it. One special argument available to submodules is the
`name` argument, which when used in `attrsOf`, gives you the name of the
attribute the submodule is defined under.

In this case, we don't easily have access to the name from the marker
submodules label option (where we could set a `default =`). Instead we
will use the `config` section of the user submodule to set a default. We
can do so using the `lib.mkDefault` modifier, which has lower precedence
than if no modifier were used.