Initial path module
Let's introduce a new module for declaring paths on the map. We'll
import a new `path.nix` module from our `marker.nix` module.

In the path module we'll define an option for declaring paths, and will
use the same `generate.requestParams` to influence the API call to
include our defined paths