Introducing path weight option
Let's also allow some customization of path styles with a `weight` option.
As already done before, we'll declare a submodule for the path style.

While we could also directly define the style.weight option in this
case, we will use the submodule in a future change to reuse the path
style definitions.

Note how we're using a new type for this, `ints.between <lower>
<upper>`, which allows integers in the given inclusive range.